Android: To support older SDKs, don't use android.os.Build.VERSION_CODES
When running Android prior to Oreo, android.os.Build.VERSION_CODES.O
does not exist at runtime and so trying to use it produces an error.
Instead, use the API level number (26) directly.

The VERSION_CODES work in Java code because the compiler replaces them
with the numbers at build time, but that doesn't happen with JavaScript.