Android: Don't request audio focus on init
Previously audio focus was requested by the playFromFile method, which
is called during initFromFile and initFromUrl. Getting audio focus this
early means that when an AUDIOFOCUS_GAIN event is received, the track
will start playing even though the caller never requested it to play.

Now the audio focus is not requested until playing the track.

Fixes #111